Ready for hyperparameter tuning.
There are a few problems with the code that I was hoping to fix after testing a few things, but it has gotten late. Here is the changelog:

- Added an optimiser class for performing hyperparameter tuning using Optuna.
- Extended BaseOptimiser to create LSTMClassifierOptimiser
    - Uses LSTMClassifierTrainer and LSTMClassifierModel to find the best accuracy from a set of suggested hyperparameters.
- Updated the console menu in main.py
    - Check CUDA availability
    - Download NLTK data packages required for text standardisation
    - Perform hyperparameter optimisation/tuning for LSTM Tweet Classifier

NOTE: You may need to comment line 13 in data_processing_utils.py (`STOP_WORDS`) to be able to compile and run the project if you are missing the required NLTK data packages. Do this, and then navigate the console menu to download these, after which you can uncomment the line.